Add files via upload
Healthcare organizations have a high potential to become an easy target for cybercrime due to their critical and vulnerable infrastructure. The rise of digitization has led to several security challenges. Identifying these critical challenges is a very important task not only from a technical point of view but also from a legal and  managerial point of view.